requirements.txt
streamlit
PyPDF2
langchain
openai
cassio
python-dotenv
